Move inner impl to detail
